Moved gdt to x86 and made it more general.

+++ source/include/sys/arch/i586/syscalls.h
@@ -20,6 +20,8 @@
 #pragma once
 
 #include <esc/common.h>
+#include <sys/arch/x86/gdt.h>
+#include <sys/cpu.h>
 
 /* some convenience-macros */
 #define SYSC_SETERROR(stack,errorCode)	((stack)->edi = (errorCode))
@@ -30,8 +32,8 @@
 #define SYSC_NUMBER(stack)				((stack)->eax)
 #define SYSC_ARG1(stack)				((stack)->esi)
 #define SYSC_ARG2(stack)				((stack)->edi)
-#define SYSC_ARG3(stack)				(*(((uint32_t*)(stack)->getSP()) + 0))
-#define SYSC_ARG4(stack)				(*(((uint32_t*)(stack)->getSP()) + 1))
-#define SYSC_ARG5(stack)				(*(((uint32_t*)(stack)->getSP()) + 2))
-#define SYSC_ARG6(stack)				(*(((uint32_t*)(stack)->getSP()) + 3))
-#define SYSC_ARG7(stack)				(*(((uint32_t*)(stack)->getSP()) + 4))
+#define SYSC_ARG3(stack)				(*(((ulong*)(stack)->getSP()) + 0))
+#define SYSC_ARG4(stack)				(*(((ulong*)(stack)->getSP()) + 1))
+#define SYSC_ARG5(stack)				(*(((ulong*)(stack)->getSP()) + 2))
+#define SYSC_ARG6(stack)				(*(((ulong*)(stack)->getSP()) + 3))
+#define SYSC_ARG7(stack)				(*(((ulong*)(stack)->getSP()) + 4))
